requirements.txt
google-genai
langchain
#openai
langchain-community
#langchain-openai
gradio
google-generativeai
streamlit